+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Added
 
  * Added `Text::into_inner()` and related methods to get ahold of a `Text` object's underlying `Image`
+ * Added `SoundData::new()` and `Source::set_repeat()`/`Source::get_repeat()` (thanks jupart!)
 
 ## Changed